+++ VisualPinball.net2015.vcxproj
@@ -488,6 +488,7 @@ copy "$(TargetPath)" "%25VPIN_DIR%25"
     <ClCompile Include="RenderDevice.cpp" />
     <ClCompile Include="RenderTarget.cpp" />
     <ClCompile Include="Sampler.cpp" />
+    <ClCompile Include="Shader.cpp" />
     <ClCompile Include="slintf.cpp" />
     <ClCompile Include="spinner.cpp" />
     <ClCompile Include="StackTrace.cpp" />
