+++ VisualPinball.net2017.vcxproj
@@ -666,6 +666,7 @@ copy "$(TargetPath)" "%25VPIN_DIR%25"
     <ClCompile Include="RenderDevice.cpp" />
     <ClCompile Include="RenderTarget.cpp" />
     <ClCompile Include="Sampler.cpp" />
+    <ClCompile Include="Shader.cpp" />
     <ClCompile Include="slintf.cpp" />
     <ClCompile Include="spinner.cpp" />
     <ClCompile Include="StackTrace.cpp" />
