+++ cmake/CMakeLists_win-x64.txt
@@ -455,6 +455,7 @@ add_executable(vpinball WIN32
    rubber.h
    Sampler.cpp
    Sampler.h
+   Shader.cpp
    Shader.h
    slintf.cpp
    slintf.h
